# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Astar Management API",
    "description": "API for Astar Management Platform",
    "contact": {
      "name": "Astar Works"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://local",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/editor/documents/{documentId}": {
      "get": {
        "tags": [
          "Editor Documents"
        ],
        "operationId": "getDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseDocumentResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Editor Documents"
        ],
        "operationId": "updateDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseDocumentResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Editor Documents"
        ],
        "operationId": "deleteDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseDocumentDeletionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editor/documents": {
      "post": {
        "tags": [
          "Editor Documents"
        ],
        "operationId": "createDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editor/documents/{documentId}/revisions": {
      "get": {
        "tags": [
          "Editor Documents"
        ],
        "operationId": "listRevisions",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseDocumentRevisionsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "JsonElement": {},
...
      "DocumentUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255.0,
            "minLength": 1.0
          },
          "summary": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonElement"
            },
            "properties": {
              "values": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JsonElement"
                }
              },
              "empty": {
                "type": "boolean"
              },
              "size": {
                "type": "integer",
                "format": "int32"
              },
              "entries": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "$ref": "#/components/schemas/JsonElement"
                    },
                    "key": {
                      "type": "string"
                    }
                  }
                },
                "uniqueItems": true
              },
              "keys": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "uniqueItems": true
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published": {
            "type": "boolean"
          },
          "favorited": {
            "type": "boolean"
          }
        }
      },
      "DocumentMetadataResponse": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonElement"
            },
            "properties": {
              "values": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JsonElement"
                }
              },
              "empty": {
                "type": "boolean"
              },
              "size": {
                "type": "integer",
                "format": "int32"
              },
              "entries": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "$ref": "#/components/schemas/JsonElement"
                    },
                    "key": {
                      "type": "string"
                    }
                  }
                },
                "uniqueItems": true
              },
              "keys": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "uniqueItems": true
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastViewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastIndexedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "published": {
            "type": "boolean"
          },
          "favorited": {
            "type": "boolean"
          }
        }
      },
      "DocumentResponse": {
        "type": "object",
        "properties": {
          "node": {
            "$ref": "#/components/schemas/EditorNodeResponse"
          },
          "latestRevision": {
            "$ref": "#/components/schemas/DocumentRevisionResponse"
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadataResponse"
          },
          "revisionCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DocumentRevisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revisionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "titleSnapshot": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "checksum": {
            "type": "string"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EditorApiResponseDocumentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentResponse"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EditorErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "EditorNodeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "materializedPath": {
            "type": "string"
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "position": {
            "type": "number",
            "format": "double"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "updatedBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "DocumentCreateRequest": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 255.0,
            "minLength": 1.0
          },
          "summary": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonElement"
            },
            "properties": {
              "values": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JsonElement"
                }
              },
              "empty": {
                "type": "boolean"
              },
              "size": {
                "type": "integer",
                "format": "int32"
              },
              "entries": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "$ref": "#/components/schemas/JsonElement"
                    },
                    "key": {
                      "type": "string"
                    }
                  }
                },
                "uniqueItems": true
              },
              "keys": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "uniqueItems": true
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "position": {
            "type": "number",
            "format": "double"
          },
          "published": {
            "type": "boolean"
          },
          "favorited": {
            "type": "boolean"
          }
        },
        "required": [
          "title"
        ]
      },
...
      "DocumentRevisionSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revisionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "titleSnapshot": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "checksum": {
            "type": "string"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocumentRevisionsResponse": {
        "type": "object",
        "properties": {
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentRevisionSummaryResponse"
            }
          }
        }
      },
      "EditorApiResponseDocumentRevisionsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentRevisionsResponse"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DocumentDeletionResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "EditorApiResponseDocumentDeletionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentDeletionResponse"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}